Created CategoryDescription component and moved dummy data into its own folder
