+++ src/DummyData/sampleCategoryData.js
@@ -1,0 +1,16 @@
+import RecyclingIcon from '@mui/icons-material/Recycling';
+
+export const categoryData = {
+  recycle: {
+    img: RecyclingIcon,
+    description: 'This is recycle. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Consequat ac felis donec et odio pellentesque diam volutpat.'
+  },
+  compost: {
+    img: '',
+    description: 'This is compost. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Consequat ac felis donec et odio pellentesque diam volutpat.'
+  },
+  trash: {
+    img: '',
+    description: 'This is trash. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Consequat ac felis donec et odio pellentesque diam volutpat.'
+  },
+}
